+++ ci/downgrade/after_downgrade_code.sh
@@ -8,7 +8,8 @@
 #
 ########################################################################
 
-# Fix the return type in Trait
-# @see https://github.com/rectorphp/rector/issues/6962
-# @see https://github.com/GatoGraphQL/GatoGraphQL/issues/1470
-sed -i 's/function getItem($key): CacheItem/function getItem($key): \\Psr\\Cache\\CacheItemInterface/' vendor/symfony/cache/Traits/AbstractAdapterTrait.php
+# Downgrading `[#Required]` to `@required` triggers a deprecation message on Symfony:
+#   array(4) { ["type"]=> int(16384) ["message"]=> string(231) "Since symfony/dependency-injection 6.3: Relying on the "@required" annotation on method "PoP\Root\Hooks\AbstractHookSet::setInstanceManager()" is deprecated, use the "Symfony\Contracts\Service\Attribute\Required" attribute instead." ["file"]=> string(95) "/app/wordpress/wp-content/plugins/gatographql/vendor/symfony/deprecation-contracts/function.php" ["line"]=> int(25) }
+# and the "php-error" class is added on the WP dashboard, adding an annoying whiteline
+# when WP_DEBUG is true. Then, comment out triggering that deprecation
+sed -i 's/trigger_deprecation/\/\/trigger_deprecation/' vendor/symfony/dependency-injection/Compiler/AutowireRequiredMethodsPass.php