+++ layers/Engine/packages/access-control/src/Hooks/AbstractConfigurableAccessControlForDirectivesInPrivateSchemaHookSet.php
@@ -12,13 +12,6 @@ abstract class AbstractConfigurableAccessControlForDirectivesInPrivateSchemaHook
     use AccessControlConfigurableMandatoryDirectivesForDirectivesHookSetTrait;
     use ConfigurableMandatoryDirectivesForDirectivesTrait, AccessControlConfigurableMandatoryDirectivesForDirectivesTrait {
         AccessControlConfigurableMandatoryDirectivesForDirectivesTrait::getMatchingEntries insteadof ConfigurableMandatoryDirectivesForDirectivesTrait;
-        // The conflict resolutions below should not be needed, because the functions are not repeated, but it is defined just once in the same source trait
-        // However, there is a bug about, still unresolved by PHP 7.2: https://bugs.php.net/bug.php?id=63911
-        // It was resolved by PHP 7.3.9, though, but handle to add compatibility up to PHP 7.1
-        AccessControlConfigurableMandatoryDirectivesForDirectivesTrait::getEntries insteadof ConfigurableMandatoryDirectivesForDirectivesTrait;
-        AccessControlConfigurableMandatoryDirectivesForDirectivesTrait::getRequiredEntryValue insteadof ConfigurableMandatoryDirectivesForDirectivesTrait;
-        AccessControlConfigurableMandatoryDirectivesForDirectivesTrait::getFieldDirectiveResolvers insteadof ConfigurableMandatoryDirectivesForDirectivesTrait;
-        AccessControlConfigurableMandatoryDirectivesForDirectivesTrait::getFieldDirectiveResolverClasses insteadof ConfigurableMandatoryDirectivesForDirectivesTrait;
     }
 
     protected function enabled(): bool