+++ layers/Engine/packages/access-control/src/Hooks/AbstractConfigurableAccessControlForFieldsInPrivateSchemaHookSet.php
@@ -16,12 +16,6 @@ abstract class AbstractConfigurableAccessControlForFieldsInPrivateSchemaHookSet 
     use AccessControlConfigurableMandatoryDirectivesForFieldsHookSetTrait;
     use ConfigurableMandatoryDirectivesForFieldsTrait, AccessControlConfigurableMandatoryDirectivesForFieldsTrait {
         AccessControlConfigurableMandatoryDirectivesForFieldsTrait::getMatchingEntries insteadof ConfigurableMandatoryDirectivesForFieldsTrait;
-        // The conflict resolutions below should not be needed, because the functions are not repeated, but it is defined just once in the same source trait
-        // However, there is a bug about, still unresolved by PHP 7.2: https://bugs.php.net/bug.php?id=63911
-        // It was resolved by PHP 7.3.9, though, but handle to add compatibility up to PHP 7.1
-        AccessControlConfigurableMandatoryDirectivesForFieldsTrait::getEntries insteadof ConfigurableMandatoryDirectivesForFieldsTrait;
-        AccessControlConfigurableMandatoryDirectivesForFieldsTrait::getFieldNames insteadof ConfigurableMandatoryDirectivesForFieldsTrait;
-        AccessControlConfigurableMandatoryDirectivesForFieldsTrait::getEntriesByTypeAndInterfaces insteadof ConfigurableMandatoryDirectivesForFieldsTrait;
     }
 
     protected function enabled(): bool