+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -8,7 +8,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Breaking changes
 
-- Bump minimum required PHP version to 7.4
+- Bump minimum required PHP version to 7.4 (#2905)
 - Allow to include Gato GraphQL as the engine to power another standalone plugin (#2897)
 
 ### Improvements