+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -10,6 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 - Bump minimum required PHP version to 7.4 (#2905)
 - Allow to include Gato GraphQL as the engine to power another standalone plugin (#2897)
+- Renamed env var `CACHE_DIR` to `CONTAINER_CACHE_DIR` (#2923)
 
 ### Improvements