+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -17,6 +17,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added convenience class for standalone plugins (#2899)
 - Allow to fetch posts with `auto-draft` status (#2911)
 - Allow disabling the private endpoint (#2913)
+- Added field `isGutenbergEditorEnabledForCustomPostType` (#2960)
 
 ### Fixed